Refactor chat API and serve it with a path prefix
