+++ learn2rag/ui/templates/compose/pipelines/continuous.yml
@@ -120,7 +120,7 @@ services:
       DEFAULT_LOCALE: 'de'
       ENABLE_OLLAMA_API: 'False'
       ENABLE_OPENAI_API: 'True'
-      OPENAI_API_BASE_URL: http://127.0.0.1:{{ports.pipeline}}
+      OPENAI_API_BASE_URL: http://127.0.0.1:{{ports.pipeline}}/api/v1
       OPENAI_API_KEY: '0p3n-w3bu!'
       ENABLE_AUTOCOMPLETE_GENERATION: 'False'
       ENABLE_CODE_EXECUTION: 'False'
